# docs/openapi.json
{
  "components": {
    "schemas": {
      "AnalyzeAccepted": {
        "description": "202 응답. job_id는 우리 내부 jobs.id — 같은 지출의 동시 제출이 하나의 잡으로\ndedupe됐는지가 이 값으로 드러난다(멱등성 스트레스 테스트·대시보드 폴링용).\n백엔드는 자기가 발급한 jobId(external_job_id로 echo)로도 GET /v1/jobs/{id} 조회\n가능. 응답 키 형식(camelCase 여부)은 미확정이라 내부 관례(snake_case) 유지.",
        "properties": {
          "external_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Job Id"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "status": {
            "default": "queued",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "dead"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id"
        ],
        "title": "AnalyzeAccepted",
        "type": "object"
      },
...
      "ContextRefreshRequest": {
        "description": "REQ-041 컨텍스트 갱신 이벤트.",
        "properties": {
          "change_type": {
            "enum": [
              "rule",
              "category",
              "params"
            ],
            "title": "Change Type",
            "type": "string"
          },
          "team_id": {
            "description": "양의 백엔드 BIGINT 기본키",
            "format": "int64",
            "title": "Team Id",
            "type": "integer"
          }
        },
        "required": [
          "team_id",
          "change_type"
        ],
        "title": "ContextRefreshRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobStatusResponse": {
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "dead"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status",
          "attempts",
          "created_at",
          "updated_at"
        ],
        "title": "JobStatusResponse",
        "type": "object"
      },
      "PolicyDraft": {
        "description": "[팀 결정 2026-07-09] 예산 카테고리 배분(budget_plan) 제거 —\n예산 현황은 지난 지출 내역 기반(ReportWriter)으로 표시.\n\n[개정 2026-08-05] policy_params(승인 정책 제안) 제거 — 기준 금액은 사용자 입력이\n원천이라 LLM이 되돌려 줄 값이 없다 (API 명세서 개정안 §1-4).",
        "properties": {
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "recommended_categories": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Recommended Categories",
            "type": "array"
          },
          "rules": {
            "items": {
              "type": "string"
            },
            "title": "Rules",
            "type": "array"
          }
        },
        "required": [
          "rules"
        ],
        "title": "PolicyDraft",
        "type": "object"
      },
      "PolicyDraftRequest": {
        "description": "마법사 1~3단계 통합 요청 (API 명세서 개정안 §1 — 회의 2·3·5번).\n\n0단계(모임 생성)는 백엔드 별도 API — 여기는 그 결과인 team_id를 받는다.\n1단계 회비(dues) · 2단계 기준 금액(force_escalation_amount) · 3단계 회칙\n(rule_source·rule_text·rule_file_ref)을 한 번에 담는다.",
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "dues": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "1인당 회비. None·0 = 없음",
            "title": "Dues"
          },
          "force_escalation_amount": {
            "description": "원 단위 기준 금액. 0 = 전건 관리자 확인",
            "minimum": 0.0,
            "title": "Force Escalation Amount",
            "type": "integer"
          },
          "initial_budget": {
            "description": "원 단위 총액",
            "exclusiveMinimum": 0.0,
            "title": "Initial Budget",
            "type": "integer"
          },
          "member_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Count"
          },
          "rule_file_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule File Ref"
          },
          "rule_source": {
            "enum": [
              "file",
              "manual",
              "ai",
              "skip"
            ],
            "title": "Rule Source",
            "type": "string"
          },
          "rule_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Text"
          },
          "team_id": {
            "description": "양의 백엔드 BIGINT 기본키",
            "format": "int64",
            "title": "Team Id",
            "type": "integer"
          },
          "team_name": {
            "title": "Team Name",
            "type": "string"
          },
          "team_type": {
            "description": "모임 유형. 백엔드 ENUM 언더바 표기('동아리_학생회')도 받는다 — 수신 시 내부 표기로 변환",
            "enum": [
              "동아리/학생회",
              "스터디",
              "친목",
              "동호회",
              "회사"
            ],
            "title": "Team Type",
            "type": "string"
          }
        },
        "required": [
          "team_id",
          "team_type",
          "team_name",
          "initial_budget",
          "force_escalation_amount",
          "rule_source"
        ],
        "title": "PolicyDraftRequest",
        "type": "object"
      },
...
      "PolicyProposal": {
        "description": "저장된 회칙 초안. 승인·거절은 PATCH /v1/proposals/{proposalId}로 기록한다.",
        "properties": {
          "draft": {
            "$ref": "#/components/schemas/PolicyDraft"
          },
          "proposal_id": {
            "title": "Proposal Id",
            "type": "string"
          },
          "reused": {
            "default": false,
            "title": "Reused",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "team_id": {
            "description": "양의 백엔드 BIGINT 기본키",
            "format": "int64",
            "title": "Team Id",
            "type": "integer"
          }
        },
        "required": [
          "proposal_id",
          "team_id",
          "status",
          "draft"
        ],
        "title": "PolicyProposal",
        "type": "object"
      },
      "PolicyProposalRequest": {
        "description": "회칙·정책 관리 화면의 초안 요청 (풀스택 협의 2026-08-04 7번).\n\n입력은 마법사 3단계와 같다 — 초안을 만드는 재료가 달라질 이유가 없다. 다른 것은\n결과를 `proposals`에 남긴다는 점과, 이미 있으면 재사용한다는 점이다.",
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "dues": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "1인당 회비. None·0 = 없음",
            "title": "Dues"
          },
          "force_escalation_amount": {
            "description": "원 단위 기준 금액. 0 = 전건 관리자 확인",
            "minimum": 0.0,
            "title": "Force Escalation Amount",
            "type": "integer"
          },
          "initial_budget": {
            "description": "원 단위 총액",
            "exclusiveMinimum": 0.0,
            "title": "Initial Budget",
            "type": "integer"
          },
          "member_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Count"
          },
          "regenerate": {
            "default": false,
            "title": "Regenerate",
            "type": "boolean"
          },
          "rule_file_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule File Ref"
          },
          "rule_source": {
            "const": "ai",
            "default": "ai",
            "title": "Rule Source",
            "type": "string"
          },
          "rule_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Text"
          },
          "team_id": {
            "description": "양의 백엔드 BIGINT 기본키",
            "format": "int64",
            "title": "Team Id",
            "type": "integer"
          },
          "team_name": {
            "title": "Team Name",
            "type": "string"
          },
          "team_type": {
            "description": "모임 유형. 백엔드 ENUM 언더바 표기('동아리_학생회')도 받는다 — 수신 시 내부 표기로 변환",
            "enum": [
              "동아리/학생회",
              "스터디",
              "친목",
              "동호회",
              "회사"
            ],
            "title": "Team Type",
            "type": "string"
          }
        },
        "required": [
          "team_id",
          "team_type",
          "team_name",
          "initial_budget",
          "force_escalation_amount"
        ],
        "title": "PolicyProposalRequest",
        "type": "object"
      },
...
      "ProposalBudgetRequest": {
        "properties": {
          "period": {
            "anyOf": [
              {
                "pattern": "^[1-9]\\d{3}-(0[1-9]|1[0-2])$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period"
          },
          "team_id": {
            "description": "양의 백엔드 BIGINT 기본키",
            "format": "int64",
            "title": "Team Id",
            "type": "integer"
          }
        },
        "required": [
          "team_id"
        ],
        "title": "ProposalBudgetRequest",
        "type": "object"
      },
      "ProposalOut": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "decided_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "team_id": {
            "description": "양의 백엔드 BIGINT 기본키",
            "format": "int64",
            "title": "Team Id",
            "type": "integer"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "team_id",
          "type",
          "payload",
          "status",
          "created_at"
        ],
        "title": "ProposalOut",
        "type": "object"
      },
      "ProposalPatch": {
        "description": "관리자 결정 기록 (§9 수락률 지표의 기록 경로).",
        "properties": {
          "decided_by": {
            "title": "Decided By",
            "type": "string"
          },
          "status": {
            "enum": [
              "accepted",
              "dismissed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "decided_by"
        ],
        "title": "ProposalPatch",
        "type": "object"
      },
...
      "RuleAmendmentRequest": {
        "properties": {
          "team_id": {
            "description": "양의 백엔드 BIGINT 기본키",
            "format": "int64",
            "title": "Team Id",
            "type": "integer"
          }
        },
        "required": [
          "team_id"
        ],
        "title": "RuleAmendmentRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ServiceToken": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "지출 자동 심사 에이전트 서버 — 잡 등록·상태 조회 (그래프 실행은 llm-worker)",
    "title": "BudgetOps LLM API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/context/refresh": {
      "post": {
        "description": "회칙 변경을 알린다. 접수만 하고 인덱싱은 비동기로 처리한다.\n\n반환된 `job_id`로 `GET /v1/jobs/{job_id}`를 조회하면 인덱싱 성공 여부를 볼 수 있고,\n아래 `/context/status`로 팀 단위 현재 상태를 볼 수도 있다.",
        "operationId": "refresh_context_v1_context_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContextRefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Context",
        "tags": [
          "context"
        ]
      }
    },
...
    "/v1/jobs/{job_id}": {
      "get": {
        "description": "심사·생성 잡의 진행 상태와 결과를 조회한다.\n\n콜백이 유실됐을 때의 **폴링 안전망**(§7.1). `job_id`는 우리 내부 id와 백엔드가\n발급한 jobId **둘 다** 받는다.\n\n`status`: `queued` → `running` → `succeeded` / `failed` / `dead`.\n`succeeded`면 `result`에 심사 결과(verdict·reasons·opinions 등)가 담긴다.\n`dead`(재시도 소진 최종 실패)면 `result`가\n`{\"error\": \"<예외 클래스명>\", \"message\": \"<실패 사유>\"}`로 채워진다 —\n`message`는 회칙 파싱 실패(`DocumentParseError` 계열)일 때만 원문이고, 그 외\n예외는 정형 문구라 관리자에게 그대로 보여줘도 된다.",
        "operationId": "read_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "잡 상태·결과 조회 (콜백 안전망)",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/policy-draft": {
      "post": {
        "description": "마법사 1~3단계 통합 요청 (LLM-005 전면 개정) — 저장 없이 결과만 돌려준다.\n\n`rule_source=ai`면 'AI 초안' 버튼 시점에 호출되어 `rules`에 회칙 초안이 담기고,\n그 외(file·manual·skip)는 '설정 완료' 시점 1회 호출로 `rules`가 빈 배열이다.\n승인 정책은 제안하지 않는다 — 기준 금액(`force_escalation_amount`)은 사용자\n입력이 원천이며 저장은 백엔드 team_settings 소관이다 (개정안 §1-3).",
        "operationId": "create_policy_draft_v1_policy_draft_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyDraft"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Policy Draft",
        "tags": [
          "drafts"
        ]
      }
    },
...
    "/v1/policy-proposals": {
      "get": {
        "description": "아직 결정되지 않은 회칙 초안을 돌려준다. 없으면 `null`이다.\n\n회의록의 \"생성된 메시지가 있으면 반환, 없으면 없다고 반환\"이 이 동작이다.\n404가 아니라 200 + `null`인 이유: '초안이 아직 없음'은 오류가 아니라 정상 상태이고,\n화면은 그때 \"AI 초안 생성하기\" 버튼을 보여주면 된다.",
        "operationId": "read_policy_proposal_v1_policy_proposals_get",
        "parameters": [
          {
            "in": "query",
            "name": "team_id",
            "required": true,
            "schema": {
              "description": "양의 백엔드 BIGINT 기본키",
              "format": "int64",
              "title": "Team Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/PolicyProposal"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Read Policy Proposal V1 Policy Proposals Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "저장된 회칙 초안 조회 (없으면 null)",
        "tags": [
          "drafts"
        ]
      },
      "post": {
        "description": "회칙 초안을 만들어 `proposals`에 남기고 돌려준다.\n\n아직 결정되지 않은 초안이 이미 있으면 **새로 만들지 않고 그것을 돌려준다**\n(`reused=true`). 관리자가 버튼을 다시 눌렀을 때 회칙이 바뀌면 안 되기 때문이다.\n일부러 새로 만들려면 `regenerate=true`를 주시면 된다 — 그때는 이전 초안을 그대로\n두고 새 제안을 하나 더 만든다(이력이 남는다).\n\n관리자가 확인한 뒤에는 `PATCH /v1/proposals/{proposalId}`로 `accepted`·`dismissed`를\n기록해 주시면 된다. 승인된 회칙을 백엔드가 저장하신 다음에는\n`POST /v1/context/refresh`를 보내 주셔야 심사에 반영된다 — 그 알림이 없으면\n회칙을 등록해도 AI는 계속 모르는 상태다.",
        "operationId": "create_policy_proposal_v1_policy_proposals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyProposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyProposal"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "회칙 초안 생성·저장 (회칙·정책 관리 화면)",
        "tags": [
          "drafts"
        ]
      }
    },
...
    "/v1/proposals": {
      "get": {
        "description": "AI가 올린 제안 목록을 돌려준다. 없으면 빈 배열이다 — 404가 아니다 (LLM-015).\n\n`type` 미지정 시 마법사 회칙 초안(`rule_draft`)은 기본 제외한다. 그쪽은\n`GET /v1/policy-proposals`(LLM-019)가 다루는 별개 자원이라, 관리자 제안함\n목록에 섞이면 같은 초안이 두 화면에 중복 노출된다. `type=rule_draft`를\n명시하면 그때는 포함된다.",
        "operationId": "read_proposals_v1_proposals_get",
        "parameters": [
          {
            "in": "query",
            "name": "team_id",
            "required": true,
            "schema": {
              "description": "양의 백엔드 BIGINT 기본키",
              "format": "int64",
              "title": "Team Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProposalOut"
                  },
                  "title": "Response Read Proposals V1 Proposals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "제안 목록 조회 (LLM-015, type 미지정 시 마법사 회칙 초안 rule_draft 제외)",
        "tags": [
          "proposals"
        ]
      }
    },
    "/v1/proposals/budget": {
      "post": {
        "description": "예산관리 페이지 AI 메시지 생성 잡 접수 (LLM-016).\n\n결과는 `GET /v1/jobs/{job_id}` 또는 `GET /v1/proposals?type=budget`으로 받는다.\npayload는 Figma 확정 3블록(category_analysis·budget_status_analysis·recommendation)\n+ figures + verified다.\n\n`verified=false`는 \"본문 수치와 집계값의 대조가 통과하지 못했다\"는 뜻이지 \"쓸 수 없는\n값\"이라는 뜻이 아니다 — 그때는 집계로 조립한 안전한 문장으로 교체해 저장하므로\n**3블록은 언제나 그대로 화면에 띄워도 된다.**",
        "operationId": "create_budget_proposal_job_v1_proposals_budget_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalBudgetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Budget Proposal Job",
        "tags": [
          "proposals"
        ]
      }
    },
    "/v1/proposals/rule-amendment": {
      "post": {
        "operationId": "create_rule_amendment_job_v1_proposals_rule_amendment_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleAmendmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Rule Amendment Job",
        "tags": [
          "proposals"
        ]
      }
    },
    "/v1/proposals/{proposal_id}": {
      "patch": {
        "operationId": "decide_proposal_v1_proposals__proposal_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Decide Proposal V1 Proposals  Proposal Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide Proposal",
        "tags": [
          "proposals"
        ]
      }
    },
...
  },
  "security": [
    {
      "ServiceToken": []
    }
  ]
}